+++ Sources/MolecularRenderer/Image/Render/RenderArgs.swift
@@ -2,7 +2,7 @@ struct RenderArgs {
   var jitterOffset: SIMD2<Float> = .zero
   var frameSeed: UInt32 = .zero
   var upscaleFactor: Float = .zero
-  var secondaryRayCount: UInt32 = .zero
+  var secondaryRayCount: Float = .zero
   var criticalPixelCount: Float = .zero
   
   static var shaderDeclaration: String {
@@ -11,7 +11,7 @@ struct RenderArgs {
       float2 jitterOffset;
       uint frameSeed;
       float upscaleFactor;
-      uint secondaryRayCount;
+      float secondaryRayCount;
       float criticalPixelCount;
     };
     """
